% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_uniprot.R
\name{ask_uniprot}
\alias{ask_uniprot}
\title{fetch known subcellular location from UniprotKB based on uniprot ids}
\usage{
ask_uniprot(uniprotID_list)
}
\arguments{
\item{UniprotID_list}{a list of UniprotIDs}
}
\value{
a list of locations
}
\description{
fetch known subcellular location from UniprotKB based on uniprot ids
}
\examples{
id_list <- c('P39864', 'D0N4E2', 'Q5BUB4', 'D0N381', 'B1NNT7', 'D0NP26')
known_locations <- ask_uniprot(id_list)
}
