% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate with the gmail API.}
\usage{
authenticate(config_path)
}
\arguments{
\item{config_path}{file path to the \code{secretsanta} config JSON.}
}
\description{
Simple convenience function to authenticate with gmail
prior to sending e-mails.  This can be skipped if the user
chooses to authenticate separately.
}
