% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CBSResult-class}
\alias{CBSResult-class}
\alias{CBSResult}
\title{An S4 class to represent minimal basic structure of CBS output objects}
\description{
An S4 class to represent minimal basic structure of CBS output objects
}
\section{Slots}{

\describe{
\item{\code{in_fasta}}{initial fasta file}

\item{\code{out_fasta}}{output fasta with only positive candidates, \cr
i.e those that passed tool filters}
}}

