% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_khdel.R
\name{check_khdel}
\alias{check_khdel}
\title{check_khdel function}
\usage{
check_khdel(input_obj, run_mode)
}
\arguments{
\item{input_obj}{input object of CBSResult superclass}

\item{run_mode}{'starter' or 'piper'}
}
\description{
This function checks presence of terminal KDEL/HDEL sequences in the candidate secreted proteins
}
\examples{
check_khdel(step1_sp2, run_mode = 'starter')
}
