% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalp_parallel.R
\name{combine_CBSResult}
\alias{combine_CBSResult}
\title{combine_CBSResult function}
\usage{
combine_CBSResult(...)
}
\arguments{
\item{arguments}{- a list of CBSResult objects to be combined}
}
\description{
Minimal function to combine objects of CBSResult class.
}
\examples{
inp2 <- CBSResult(in_fasta = readAAStringSet(system.file("extdata", "tail_prot.fasta", package = "SecretSanta")))
inp3 <- CBSResult(in_fasta = readAAStringSet(system.file("extdata", "tail2_prot.fasta", package = "SecretSanta")))
inp4 <- CBSResult(in_fasta = readAAStringSet(system.file("extdata", "sample_prot_100.fasta", package = "SecretSanta")))
combined_CBS <- combine_CBSResult(inp2, inp3, inp4)
}
