% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalp_parallel.R
\name{combine_SignalpResult}
\alias{combine_SignalpResult}
\title{This function combines multiple instances of SignalpResult class, typically generated with parLapply}
\usage{
combine_SignalpResult(arguments)
}
\arguments{
\item{arguments}{- a list of SignalpResult objects to be combined in one}
}
\description{
This function combines multiple instances of SignalpResult class, typically generated with parLapply
}
\examples{
my_pa <- manage_paths(system.file(
                      "extdata",
                      "sample_paths",
                      package = "SecretSanta"
                      ))
                                                 
inp2 <- CBSResult(in_fasta =
                  readAAStringSet(
                  system.file("extdata",
                  "tail_prot.fasta",
                  package = "SecretSanta"
                 )
                  ))
inp3 <- CBSResult(in_fasta = 
                  readAAStringSet(
                  system.file("extdata",
                  "tail2_prot.fasta",
                  package = "SecretSanta"
                 )
                  ))
inp4 <- CBSResult(in_fasta =
                 readAAStringSet(
                 system.file("extdata",
                 "sample_prot_100.fasta",
                 package = "SecretSanta"
                )
                 ))
                  
sp1 <- signalp(input_obj = inp2,
               version = 2,
               organism_type = 'euk',
               run_mode = 'starter',
               paths = my_pa)
sp2 <- signalp(input_obj = inp3,
               version = 2,
               organism_type = 'euk',
               run_mode = 'starter',
               paths = my_pa)
sp3 <- signalp(input_obj = inp4,
               version = 2,
               organism_type = 'euk',
               run_mode = 'starter', 
               paths = my_pa)
obj <- list(sp1, sp2, sp3)
combined_sp <- combine_SignalpResult(obj)
}
