% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetp_parallel.R
\name{combine_TargetpResult}
\alias{combine_TargetpResult}
\title{This helper function combines multiple instances of TargetpResult class, typically generated with parLapply}
\usage{
combine_TargetpResult(arguments)
}
\arguments{
\item{arguments}{- a list of TargetpResult objects to be combined in one}
}
\description{
This helper function combines multiple instances of TargetpResult class, typically generated with parLapply
}
\examples{
inp2 <- CBSResult(in_fasta = readAAStringSet(system.file("extdata", "tail_prot.fasta", package = "SecretSanta")))
inp3 <- CBSResult(in_fasta = readAAStringSet(system.file("extdata", "tail2_prot.fasta", package = "SecretSanta")))
inp4 <- CBSResult(in_fasta = readAAStringSet(system.file("extdata", "sample_prot_100.fasta", package = "SecretSanta")))

tp1 <- targetp(input_obj = inp2, network_type = 'N', run_mode = 'starter', paths = my_pa)
tp2 <- targetp(input_obj = inp3, network_type = 'N', run_mode = 'starter', paths = my_pa)
tp3 <- targetp(input_obj = inp4, network_type = 'N', run_mode = 'starter', paths = my_pa)

obj <- list(tp1, tp2, tp3)
}
