% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmhmm_parallel.R
\name{combine_TMhmmResult}
\alias{combine_TMhmmResult}
\title{This function combines multiple instances of TMhmmResult class, typically generated with parLapply}
\usage{
combine_TMhmmResult(arguments)
}
\arguments{
\item{arguments}{- a list of TMhmmResult objects to be combined in one}
}
\description{
This function combines multiple instances of TMhmmResult class, typically generated with parLapply
}
\examples{
inp2 <- CBSResult(in_fasta = readAAStringSet(system.file("extdata", "tail_prot.fasta", package = "SecretSanta")))
inp4 <- CBSResult(in_fasta = readAAStringSet(system.file("extdata", "sample_prot_100.fasta", package = "SecretSanta")))

sp1 <- signalp(input_obj = inp2, version = 2, organism_type = 'euk', run_mode = 'starter', paths = my_pa)
sp3 <- signalp(input_obj = inp4, version = 2, organism_type = 'euk', run_mode = 'starter', paths = my_pa)

tm1 <- tmhmm(sp1, paths = my_pa, TM = 1)
tm3 <- tmhmm(sp3, paths = my_pa, TM = 1)
obj <- list(tm1, tm3)
 combined_tm <- combine_TMhmmResult(obj)
}
