% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetp_parallel.R
\name{combine_TpResult}
\alias{combine_TpResult}
\title{combine_TpResult function}
\usage{
combine_TpResult(arguments)
}
\arguments{
\item{arguments}{- a list of TargetpResult objects to be combined.}
}
\value{
TargetpResult object
}
\description{
This function combines multiple instances of TargetpResult class,
typically generated with parLapply when running targetp in a paralle mode.
}
\examples{

inp2 <- CBSResult(in_fasta = readAAStringSet(
                             system.file("extdata",
                             "tail_prot.fasta",
                             package = "SecretSanta")))
inp3 <- CBSResult(in_fasta = readAAStringSet(
                             system.file("extdata",
                             "tail2_prot.fasta",
                             package = "SecretSanta")))
inp4 <- CBSResult(in_fasta = readAAStringSet(
                             system.file("extdata",
                             "sample_prot_100.fasta",
                             package = "SecretSanta")))

tp1 <- targetp(input_obj = inp2,
               network = 'N',
               run_mode = 'starter',
               paths = my_pa)
tp2 <- targetp(input_obj = inp3,
               network = 'N',
               run_mode = 'starter',
               paths = my_pa)
tp3 <- targetp(input_obj = inp4,
               network = 'N',
               run_mode = 'starter',
               paths = my_pa)

obj <- list(tp1, tp2, tp3)
 combined_tp <- combine_TpResult(obj)
}
