% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{crop_names}
\alias{crop_names}
\title{crop_names function}
\usage{
crop_names(x)
}
\value{
character vector
}
\description{
helper function crop long names for AAStringSet object, return
character vector
}
\keyword{internal}
