% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ErResult-class}
\alias{ErResult-class}
\alias{ErResult}
\title{An S4 class to represent intermediate and final outputs of the ER motif checks (KDEL/HDEL)}
\description{
An S4 class to represent intermediate and final outputs of the ER motif checks (KDEL/HDEL)
}
\section{Slots}{

\describe{
\item{\code{retained}}{sequences containing terminal ER-retention signals}

\item{\code{mootifs}}{dummy slot}
}}

