% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ErResult-class}
\alias{ErResult-class}
\alias{ErResult}
\alias{getRetainedfasta}
\alias{getRetainedfasta,ErResult-method}
\title{accessor functions for objects of ErResult S4 class, outputs of the
check_khedel fucntion}
\usage{
getRetainedfasta(theObject)

\S4method{getRetainedfasta}{ErResult}(theObject)
}
\arguments{
\item{theObject}{an object of WolfResult class}

\item{retained_fasta}{sequences with ER retention signals}
}
\description{
accessor functions for objects of ErResult S4 class, outputs of the
check_khedel fucntion

accessors for ErResult objects
}
\section{Slots}{

\describe{
\item{\code{retained_fasta}}{sequences with ER retention signals}
}}

\examples{
aa <- readAAStringSet(system.file("extdata","small_prot.fasta",
package = "SecretSanta"), use.names = TRUE)
er <- check_khdel(CBSResult(in_fasta = aa), run_mode = 'starter')
class(er)
}
