% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_function.R
\name{hi_function}
\alias{hi_function}
\title{Hi Function}
\usage{
hi_function(user_name)
}
\arguments{
\item{name}{enter your user name here}
}
\description{
This is a test fucntion
}
\examples{
hi_function()
}
\keyword{hi}
