% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-slicer.R
\name{m_slicer}
\alias{m_slicer}
\title{m_slicer function}
\usage{
m_slicer(input_object, length_threshold)
}
\arguments{
\item{input_object}{an instance of CBSResult class or AAStringSet class containing protein sequences as on of the     attributes}

\item{len_threshold}{sliced sequences below this threshold will be discarded}
}
\description{
Experimental option
This function generates all possible subsequences starting with M.
Assumption: translation start sites might be mis-predicted in the original set of proteins.
Output of this step can be used as an iput for secretome prediction pipeline
}
\examples{
aa <- readAAStringSet(system.file("extdata", "sample_prot.fasta", package = "SecretSanta"), use.names = TRUE)
m_slicer(aa, 100)
}
