% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-santa-data.R
\name{make_santa_data}
\alias{make_santa_data}
\title{Make secret santa data}
\usage{
make_santa_data(config)
}
\arguments{
\item{config}{the parsed \code{secretsanta} config obtained by passing the
config JSON file through \code{jsonlite::fromJSON}.}
}
\description{
This function turns the config data into secret santa
assignments and composes the customized e-mail message.
This function is called from \code{run_secret_santa} but is
exported in case the user wishes to study how the package
makes assignments and build messages.
}
