% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_slicer.R
\name{my_slicer}
\alias{my_slicer}
\title{my_slicer function}
\usage{
my_slicer(input_object, length_threshold, run_mode)
}
\arguments{
\item{input_object}{an instance of CBSResult class or AAStringSet class containing protein sequences as on of the     attributes}

\item{run_mode}{\itemize{
\item  \strong{slice} - to just slice input fasta, regardless of it's origin;
\item  \strong{rescue} - to get proteins not predicted to be secreted on the initial run, generate slices; 
  }}

\item{len_threshold}{sliced sequences below this threshold will be discarded}
}
\description{
Experimental option.
This function generates all possible subsequences starting with M.
Assumption: translation start sites might be mis-predicted in the original set of proteins.
Output of this step can be used as an input for secretome prediction pipeline to rescue secreted proteins
with mis-predicted start sites.
}
