% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_signalp.R
\name{parse_signalp}
\alias{parse_signalp}
\title{parse_signalp function}
\usage{
parse_signalp(data_path)
}
\arguments{
\item{data_path}{text output from signalp2 or signalp3, test: read from the file}
}
\description{
This function parses signalp2 and signalp3 output
}
\examples{
res2 <- parse_signalp("SecretSanta/inst/extdata/sample_prot_signalp2_out")
}
