% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived_functions.R
\name{Parse_signalP_web}
\alias{Parse_signalP_web}
\title{Parse_signalP_web function}
\usage{
Parse_signalP_web(url)
}
\arguments{
\item{url}{url used to produce the output
Parse_signalP_web()}
}
\description{
This function scraps web output of SignalP tool and returns output in short format organised in a dataframe
}
