% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_signalp.R
\name{pipe_signalp}
\alias{pipe_signalp}
\title{pipe_signalp function}
\usage{
pipe_signalp(piping_sequence)
}
\arguments{
\item{piping_sequence}{specify piping sequence: what to pipe where and when}
}
\description{
This function pipes outputs from different version of signalp
}
\examples{


}
