% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSP.R
\name{predictSP}
\alias{predictSP}
\title{PredictSP function}
\usage{
predictSP(proteins, version)
}
\arguments{
\item{proteins}{input file/data frame with proteins}

\item{version}{spcify SignalP version to run}
}
\description{
This function call SignalP for prediction of signal peptides for amino acid sequences
}
\examples{
predictSP()
}
\keyword{HMM}
\keyword{peptide}
\keyword{peptide,}
\keyword{prediction,}
\keyword{proteins,}
\keyword{signal}
