% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultCookie_objects.R
\docType{class}
\name{resultCookie-class}
\alias{resultCookie-class}
\alias{resultCookie}
\title{An S4 class to represent intermediate and final outputs of the
secretome prediction pipeline}
\description{
An S4 class to represent intermediate and final outputs of the
secretome prediction pipeline
}
\section{Slots}{

\describe{
\item{\code{paths}}{paths to CBS exceutibles, do I need it here?}

\item{\code{fasta}}{initial fasta file}

\item{\code{mature_fasta}}{fasta with mature sequences}

\item{\code{sp_tibble}}{standard tibble with outputs obtained from signalp}

\item{\code{tmhmm_tibble}}{tmhmm results}

\item{\code{targetp_tibble}}{targetp results}

\item{\code{er_motifs}}{ER motifs: (K/H)DEL}
}}

