% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-secret-santa.R
\name{run_secret_santa}
\alias{run_secret_santa}
\title{Run entire secret santa process}
\usage{
run_secret_santa(config_path)
}
\arguments{
\item{config_path}{file path to the \code{secretsanta} config JSON.}
}
\description{
Caution!  This function will send e-mails with
secret santa assignments!  This function executes the entire
process from making assignments to sending e-mails to
participants informing them of their match.  It also writes
a log file of givers and recipients as a CSV.
}
