% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_factory.R
\name{secret_factory}
\alias{secret_factory}
\title{generate proteins with alternative translation start sites
turn individual function generation into function factory?
write a basic function to generate a generic wrapper for 
a command line tool}
\usage{
secret_factory()
}
\arguments{
\item{method}{we want to use}

\item{specs}{for input}

\item{paths}{where to get the tool}
}
\description{
generate proteins with alternative translation start sites
turn individual function generation into function factory?
write a basic function to generate a generic wrapper for 
a command line tool
}
