\name{secretsanta}
\alias{secretsanta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Secret Santa List Generator
}
\description{
Takes as input a vector of friends, family, cowrokers, randos, etc. and returns a data.frame with a recipient.
}
\usage{
secretsanta(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{names}{input character vector}
  \item{sean}{boolean: Should we use Sean's method or not}

}
\details{
   jrwasjdfglkv ajs fgpo wqrghwaosj gv asidpgkaws
}
\value{
Returns a data frame of secret santa list.
}
\references{
Kerr, 2024 "Ho! Secret Santa!" Journal of Holiday Computer Science.
}
\author{
Kerr, S. P.
}
\note{
Bart was late by 13 minutes. Last day of class.  Can you believe it?

}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
asdhjgajsdhfg
}
\examples{
## The function is currently defined as
secretsanta(c("Leah","Leigha","Lea","Leeya"))
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
