% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_signalp.R
\name{signalp}
\alias{signalp}
\title{signalp function}
\usage{
signalp(input_obj, version, organism_type)
}
\arguments{
\item{version}{signalp version to run, allowed  values: 2, 3, 4, 4.1}

\item{organism_type}{Allowed values: 'euk', 'gram+', 'gram-'}

\item{input}{input object (any from SecretSanta objects family) with protein sequences as on of the attributes}
}
\description{
This function calls local SignalP
Please ensure that respective version of SignalP is downloaded, installed and respective path is added to $PATH variable
}
\examples{
result <- signalp(proteins = "SecretSanta/inst/extdata/sample_prot.fasta", organism_type = 'euk', version = 4) #old
w <- SignalpResult()
class(w)
w <- setInfasta(w, bb) #instance of the SignalpResult object with populated Infasta attribute
getInfasta(w)
r4 <- signalp(w, version = 4, 'euk')
r2 <- signalp(w, version = 3, 'euk')
r2 <- signalp(w, version = 4, organism_type = 'gra')
}
