% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalp_parallel.R
\name{signalp_parallel}
\alias{signalp_parallel}
\title{signalp_parallel function}
\usage{
signalp_parallel(input_obj, version, organism_type, run_mode, paths,
  truncate = NULL)
}
\arguments{
\item{version}{signalp version to run, supported versions:
\itemize{
\item 2
\item 3
\item 4
\item 4.1
}}

\item{organism_type}{possible values: 
\itemize{
\item euk - for eukaryotes;
\item gram+ - for gram-positive bacteria;
\item gram- - for gram-negative bacteria;
}}

\item{run_mode}{\itemize{
\item starter - if it is the first step in pipeline;
\item piper - if you run this function on the output of other CBS tools;
}}

\item{paths}{tibble with paths to external dependencies, generated with \code{\link{manage_paths}} function}

\item{truncate}{logical, if TRUE - sequences longer 2000 residues will be truncated to this length limit and renamed. If FALSE - long sequences will be excluded from the analysis. Default = TRUE.}

\item{input_object}{an instance of CBSResult class containing protein sequences as on of the attributes}
}
\value{
an object of SignalpResult class
}
\description{
This function calls local signalp to predict the presence and location of signal peptide cleavage sites in amino acid sequences; automatically splits large input files (>500 sequnces) and runs signalp prediction as an embarassingly parallel process on all the CPUs available.
}
