% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SignalpResult-class}
\alias{SignalpResult-class}
\alias{SignalpResult}
\title{An S4 class to represent intermediate and final outputs of the signalp prediction step}
\description{
An S4 class to represent intermediate and final outputs of the signalp prediction step
}
\section{Slots}{

\describe{
\item{\code{in_fasta}}{initial fasta file, do we need to drag it along?}

\item{\code{out_fasta}}{output fasta with only positive candidates, i.e sequences predicted to be secreted at this step}

\item{\code{mature_fasta}}{fasta with mature sequences}

\item{\code{sp_version}}{version of signalp used to generate this object}

\item{\code{sp_tibble}}{Object of class tibble, contains ... columns:
\itemize{
  \item gene_id - unique id of the sequence
  \item Cmax - 
  \item Cpos - 
  \item Ymax - 
  \item Ypos - 
  \item Smax -
  \item Spos -
  \item Smean -
  \item Prediction -
  \item Dmaxcut -
  \item Networks used -
  }}
}}

