% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_parallel.R
\name{split_XStringSet}
\alias{split_XStringSet}
\title{split_XStringSet function}
\usage{
split_XStringSet(string_set, chunk_size)
}
\description{
This function splits large XStringSet files into chunks of given size and writes them in tmp files
}
\examples{
large_aa <- readAAStringSet(system.file("extdata", "Ppalm_prot_ALI_PLTG.fasta", package = "SecretSanta"))
split_XStringSet(large_aa, 1000)
res <- split_XStringSet(large_aa, 1000)
}
