% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_signalp.R
\name{Submit_signalP_web}
\alias{Submit_signalP_web}
\title{Submit_signalP_web function}
\usage{
Submit_signalP_web(fasta, mode)
}
\arguments{
\item{fasta}{fasta file with protein sequences}

\item{mode}{specify output format}
}
\description{
This function submits input to SignalP web tool and returns the actual url which could be used by Parse_signalP_web
}
\examples{
Submit_signalP_web()
}
\keyword{HMM,}
\keyword{peptide}
\keyword{peptide,}
\keyword{prediction,}
\keyword{proteins,}
\keyword{scrapping}
\keyword{signal}
\keyword{web}
