% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetp_parallel.R
\name{targetp_parallel}
\alias{targetp_parallel}
\title{targetp_parallel function}
\usage{
targetp_parallel(input_object, network_type, run_mode, paths)
}
\arguments{
\item{input_object}{an instance of CBSResult class containing protein sequences as on of the attributes}

\item{network_type}{possible values: 
\itemize{
\item P - for plants;
\item N - for non-plants;
}}

\item{run_mode}{\itemize{
\item starter - if it is the first step in pipeline;
\item piper - if you run this function on the output of other CBS tools;
}}

\item{paths}{tibble with paths to external dependencies, generated with \code{\link{manage_paths}} function}
}
\value{
an object of TargetpResult class
}
\description{
This function calls local targetp to predict subcellular localisation of a protein.
}
