% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{methods}
\name{setTPtibble}
\alias{setTPtibble}
\alias{setTPtibble,TargetpResult-method}
\alias{setTPtibble}
\alias{getTPtibble,TargetpResult-method}
\alias{getTPtibble}
\title{Accessors for TargetpResult objects}
\usage{
setTPtibble(theObject, tp_tibble)

\S4method{setTPtibble}{TargetpResult}(theObject, tp_tibble)

\S4method{getTPtibble}{TargetpResult}(theObject)
}
\arguments{
\item{theObject}{an object of TargetpResult class}

\item{tp_tibble}{parsed targetp output in tabular format}
}
\description{
Accessors for TargetpResult objects
}
