% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-secret-santa.R
\name{test_secret_santa}
\alias{test_secret_santa}
\title{Test secret santa process}
\usage{
test_secret_santa(config_path, mail_to)
}
\arguments{
\item{config_path}{file path to the \code{secretsanta} config JSON.}

\item{mail_to}{the e-mail address to which the test message will be sent.}
}
\description{
Prior to exeuting \code{run_secret_santa}, you may wish to test
the setup by sending yourself an e-mail generated by the
package.  This function will mimic the real process, but will
only send one e-mail, and only to the address you specify,
so you can ensure it looks the way you intend.
}
