% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmhmm_parallel.R
\name{tmhmm_parallel}
\alias{tmhmm_parallel}
\title{tmhmm function}
\usage{
tmhmm_parallel(input_obj, paths, TM)
}
\arguments{
\item{input_obj}{input object, an instance of CBSResult class, \cr
input should contain mature_fasta; 
in the full-length proteins N-terminal signal peptide could
be erroneously predicted as TM domain, avoid this}

\item{paths}{tibble with paths to external dependencies, generated with 
\code{\link{manage_paths}} function}

\item{TM}{allowed number of TM domains in mature peptides, 
recommended value <= 1; use 0 for strict filtering}
}
\value{
TMhmmResult object
}
\description{
This function calls local TMHMM, expects CBSresult class objects with 
populated mature_fasta slot.
To generate CBSResult objects with mature_fasta slots run preffered version 
of signalp first.
Parallel version
}
\examples{
          
my_pa <- manage_paths(system.file("extdata",
                                  "sample_paths",
                                   package = "SecretSanta"))
inp <- SignalpResult()
aa <- readAAStringSet(system.file("extdata",
                                  "sample_prot_100.fasta",
                                  package = "SecretSanta"))
inp <- setInfasta(inp, aa)
s1_sp2 <- signalp(inp,
                  version = 2,
                  'euk',
                  run_mode = "starter",
                  paths = my_pa)
tm <- tmhmm_parallel(s1_sp2, paths = my_pa, TM = 1)
tmhmm_parallel(s1_sp2, paths = my_pa, TM = 1)
}
