% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TMhmmResult-class}
\alias{TMhmmResult-class}
\alias{TMhmmResult}
\title{An S4 class to represent intermediate and final outputs of the TMHMM prediction step}
\description{
An S4 class to represent intermediate and final outputs of the TMHMM prediction step
}
\section{Slots}{

\describe{
\item{\code{in_fasta}}{initial fasta file}

\item{\code{in_mature_fasta}}{input mature fasta,}

\item{\code{out_fasta}}{output fasta with only positive full-length candidates, candidates without TM in mature sequences}

\item{\code{out_mature_fasta}}{output mature, conatins mature sequences without TM domains}

\item{\code{tm_tibble}}{tibble with outputs obtained from TMHMM}
}}

