% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topcons.R
\name{topcons}
\alias{topcons}
\title{predict transmembrane domains with TOPCONS}
\usage{
topcons()
}
\arguments{
\item{input_obj}{an instance of SignalpResult class, \cr
input should contain mature_fasta slot;}

\item{TM}{allowed number of TM domains in mature peptides,
recommended value <= 1; use TM = 0 for strict filtering}

\item{paths}{if topcons is not acessible globally, a file
conatining a full path to it's executable should be provided; for details
please check SecretSanta vignette.}
}
\value{
TopconsResult object
}
\description{
This function calls local TOPCONS to predict transmembarne domains in protein
sequence. \cr
\cr
}
