% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{truncate_seq}
\alias{truncate_seq}
\title{truncate_seq}
\usage{
truncate_seq(truncate, seq_set, threshold)
}
\arguments{
\item{truncate}{truncate or not}

\item{seq_set}{XStringSet}

\item{threshold}{length threshold}
}
\value{
truncated AAStringSet
}
\description{
helper function to truncate long sequences or throw them away, otherwise
signalp will break (at least signalp2 and signalp3 will)
}
\keyword{internal}
