% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wolfpsort.R
\name{wolfpsort}
\alias{wolfpsort}
\title{run_WolfPsort function}
\usage{
wolfpsort(input_obj, organism)
}
\arguments{
\item{input_obj}{Object of CSBResult class}

\item{organism}{set relevant taxonomic group,
Options include: plant, animal, fingi;}
}
\description{
This function runs WoLF PSORT to predict protein cellular sub-localisation
And returns the most probbale
Recommended to run on the late stages of secretome prediction
}
