% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{WolfResult-class}
\alias{WolfResult-class}
\alias{WolfResult}
\title{An S4 class to represent outputs of WolfPsort}
\description{
An S4 class to represent outputs of WolfPsort
}
\section{Slots}{

\describe{
\item{\code{wolf_tibble}}{tibble with outputs obtained from wolfpsort
\itemize{
  \item gene_id - unique sequence id
  \item localization - the most probable localization
}}
}}

