% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{methods}
\name{setWOLFtibble}
\alias{setWOLFtibble}
\alias{setWOLFtibble,WolfResult-method}
\alias{setWOLFtibble}
\alias{getWOLFtibble}
\alias{getWOLFtibble,WolfResult-method}
\alias{getWOLFtibble}
\title{Some title}
\usage{
setWOLFtibble(theObject, wolf_tibble)

\S4method{setWOLFtibble}{WolfResult}(theObject, wolf_tibble)

getWOLFtibble(theObject)

\S4method{getWOLFtibble}{WolfResult}(theObject)
}
\arguments{
\item{theObject}{some object}

\item{wolf_tibble}{wolf result}

\item{theObject}{some object}

\item{wolf_tibble}{wolf result}
}
\description{
Some title

Some title
}
