% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CBSResult-class}
\alias{CBSResult-class}
\alias{CBSResult}
\alias{setInfasta}
\alias{setInfasta,CBSResult-method}
\alias{getInfasta,CBSResult-method}
\alias{getInfasta}
\alias{setOutfasta,CBSResult-method}
\alias{setOutfasta}
\alias{getOutfasta,CBSResult-method}
\alias{getOutfasta}
\alias{getFastas,CBSResult-method}
\alias{getFastas}
\title{accessor functions for objects of \code{\link{CBSResult}} S4 class, minimal parent class for
SecretSanat outputs}
\usage{
setInfasta(theObject, in_fasta)

\S4method{setInfasta}{CBSResult}(theObject, in_fasta)

\S4method{getInfasta}{CBSResult}(theObject)

\S4method{setOutfasta}{CBSResult}(theObject, out_fasta)

\S4method{getOutfasta}{CBSResult}(theObject)

\S4method{getFastas}{CBSResult}(theObject)
}
\arguments{
\item{theObject}{\code{\link{CBSResult}} object}

\item{in_fasta}{input fasta, AAStringSet object}

\item{out_fasta}{output fasta, AAStringSet object}
}
\value{
CBSResult object
}
\description{
accessor functions for objects of \code{\link{CBSResult}} S4 class, minimal parent class for
SecretSanat outputs

accessors for CBSResult objects
}
\section{Slots}{

\describe{
\item{\code{in_fasta}}{initial fasta file}

\item{\code{out_fasta}}{output fasta file with only positive candidates,
i.e those that passed tool filters}
}}

\examples{
library(Biostrings)
# read fasta file in AAStringSet object
aa <- readAAStringSet(system.file("extdata","sample_prot_100.fasta",
                        package = "SecretSanta"))
# create an instance of CBSResult class:
cbs <- CBSResult( in_fasta = aa, out_fasta = aa)
# check that CBSResult instance is valid
validObject(cbs)

# extract in_fasta attribute
getInfasta(cbs)

# extract out_fasta attribute
getOutfasta(cbs)
}
