% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{methods}
\name{setInfasta}
\alias{setInfasta}
\alias{setInfasta,CBSResult-method}
\alias{setInfasta}
\alias{getInfasta,CBSResult-method}
\alias{getInfasta}
\alias{setOutfasta}
\alias{setOutfasta,CBSResult-method}
\alias{setOutfasta}
\alias{getOutfasta,CBSResult-method}
\alias{getOutfasta}
\title{Accessors for CBSResult objects}
\usage{
setInfasta(theObject, in_fasta)

\S4method{setInfasta}{CBSResult}(theObject, in_fasta)

\S4method{getInfasta}{CBSResult}(theObject)

setOutfasta(theObject, out_fasta)

\S4method{setOutfasta}{CBSResult}(theObject, out_fasta)

\S4method{getOutfasta}{CBSResult}(theObject)
}
\arguments{
\item{theObject}{CBSResult object}

\item{in_fasta}{input fasta, AAStringSet object}

\item{out_fasta}{output fasta, AAStringSet object}
}
\description{
Accessors for CBSResult objects
}
