% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CBSResult-class}
\alias{CBSResult-class}
\alias{CBSResult}
\title{An S4 class to represent minimal basic structure of CBS output objects.}
\description{
An S4 class to represent minimal basic structure of CBS output objects.
}
\section{Slots}{

\describe{
\item{\code{in_fasta}}{initial fasta file}

\item{\code{out_fasta}}{output fasta with only positive candidates,
i.e those that passed tool filters}
}}

\examples{
# read fasta file in AAStringSet object
aa <- readAAStringSet(system.file("extdata", "sample_prot_100.fasta", package = "SecretSanta"), use.names = TRUE)

# create an emty instance of CBSResult class
cbs <- CBSResult()

# populate in_fasta attribute with aa
cbs <- setInfasta(cbs, aa)

# populate out_fasta attribute with aa
cbs <- setOutfasta(cbs, aa)

# check that created CBSResult instance is valid
validObject(cbs)

# extract in_fasta attribute
getInfasta(cbs)

# extract out_fasta attribute
getOutfasta(cbs)
}
