% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_khdel.R
\name{check_khdel}
\alias{check_khdel}
\title{check C-terminal ER-retention signals}
\usage{
check_khdel(input_obj)
}
\arguments{
\item{input_obj}{input object of CBSResult class}
}
\value{
ErResult object
}
\description{
This function checks the presence of terminal \strong{KDEL/HDEL} motifs in 
the provided amino acid equences.
}
\examples{
# check ER retention signals in CBSResult object
# before running signalp or any other
# predictions:
aa <- readAAStringSet(system.file("extdata", "sample_prot_100.fasta",
                                    package = "SecretSanta"))
inp <- SignalpResult(in_fasta = aa[1:10])
et_s <- check_khdel(inp,
    run_mode = 'starter')

# generate signalp predictions:
step1_sp2 <- signalp(et_s, version = 4, organism = 'euk', 
    run_mode = 'starter')
# check ER retention signal in the signalp output
et_result <- check_khdel(step1_sp2)
}
