% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_khdel.R
\name{check_khdel}
\alias{check_khdel}
\title{check C-terminal ER-retention signals}
\usage{
check_khdel(input_obj, pattern = c("prosite", "elm", "strict"))
}
\arguments{
\item{input_obj}{input object of CBSResult class}

\item{pattern}{ER-pattern to check for:
\itemize{
\item \code{pattern = 'prosite'} - "[KRHQSA][DENQ]EL$>"
\item \code{pattern = 'elm'} - "[KRHQSAP][DENQT]EL$"  
\item \code{pattern = 'strict'} - "[KH]DEL$"
}}
}
\value{
ErResult object
}
\description{
This function checks the presence of terminal \strong{KDEL/HDEL} motifs in 
the provided amino acid equences.
}
\examples{
# check ER retention signals in CBSResult object
# before running signalp or any other
# predictions:
aa <- readAAStringSet(system.file("extdata", "sample_prot_100.fasta",
                                    package = "SecretSanta"))
inp <- SignalpResult(in_fasta = aa[1:10])
# generate signalp predictions:
step1_sp2 <- signalp(inp, version = 4, organism = 'euk', 
    run_mode = 'starter')
# check ER retention signal in the signalp output, PROSITE pattern
et_result <- check_khdel(step1_sp2, pattern = 'prosite')
}
