% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_khdel.R
\name{check_khdel}
\alias{check_khdel}
\title{check_khdel function}
\usage{
check_khdel(proteins, output_type)
}
\arguments{
\item{proteins}{input file with proteins}

\item{output_type}{what we want from the output: fasta, tibble? list of ids?}

\item{input_type}{path, object or something else?}
}
\description{
This function checks for terminal KDEL/HDEL sequences in the candidate secreted proteins
}
\examples{
check_khdel("/home/anna/anna/Labjournal/SecretSanta/inst/extdata/sample_prot.fasta", 'some')
}
