% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_khdel.R
\name{check_khdel}
\alias{check_khdel}
\title{check_khdel function}
\usage{
check_khdel(input_obj, run_mode)
}
\arguments{
\item{input_obj}{input object of CBSResult superclass}
}
\description{
This function checks presence of terminal KDEL/HDEL sequences in the candidate secreted proteins
}
\examples{
check_khdel("/home/anna/anna/Labjournal/SecretSanta/inst/extdata/sample_prot.fasta", 'some')
}
