% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ErResult-class}
\alias{ErResult-class}
\alias{ErResult}
\title{An S4 class to represent intermediate and final outputs of the TMHMM prediction step}
\description{
An S4 class to represent intermediate and final outputs of the TMHMM prediction step
}
\section{Slots}{

\describe{
\item{\code{retained}}{- sequences with ER retention signals}
}}

\examples{
aa <- readAAStringSet(system.file("extdata", 
                                  "small_prot.fasta",
                                   package = "SecretSanta"),
                      use.names = TRUE)
er <- check_khdel(CBSResult(in_fasta = aa), run_mode = 'starter')
class(er)
}
