% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ErResult-class}
\alias{ErResult-class}
\alias{ErResult}
\title{An S4 class to represent intermediate and final outputs of the TMHMM prediction step}
\description{
An S4 class to represent intermediate and final outputs of the TMHMM prediction step
}
