% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ErResult-class}
\alias{ErResult-class}
\alias{ErResult}
\alias{getErRetained,ErResult-method}
\alias{getErRetained}
\title{accessor functions for objects of ErResult S4 class, outputs of the
check_khedel fucntion}
\usage{
\S4method{getErRetained}{ErResult}(theObject)
}
\description{
accessor functions for objects of ErResult S4 class, outputs of the
check_khedel fucntion
}
\section{Slots}{

\describe{
\item{\code{retained}}{- sequences with ER retention signals}
}}

\examples{
aa <- readAAStringSet(system.file("extdata","small_prot.fasta",
package = "SecretSanta"), use.names = TRUE)
er <- check_khdel(CBSResult(in_fasta = aa), run_mode = 'starter')
class(er)
}
