% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_paths.R
\name{manage_paths}
\alias{manage_paths}
\title{manage_paths function}
\usage{
manage_paths(path_file)
}
\arguments{
\item{path_file}{2-column space-separated file with listed paths for different signalp versions}
}
\description{
This function calls manages pathways for signalP versions
}
\examples{
secret_paths <- manage_paths("SecretSanta/inst/extdata/sample_paths")
}
