% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_paths.R
\name{manage_paths}
\alias{manage_paths}
\title{manage_paths function}
\usage{
manage_paths(path_file)
}
\arguments{
\item{path_file}{2-column space-separated file with listed paths for external dependencies
\itemize{
\item first column should contain tool name
\item second column should contain full path to the tool's executable
}}
}
\value{
tibble with verified file paths
}
\description{
This function reads and stores pathways for external tools used in secertome prediction pipelie\cr
}
\examples{
secret_paths <- manage_paths("SecretSanta/inst/extdata/sample_paths")
}
