% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_paths.R
\name{manage_paths}
\alias{manage_paths}
\title{manage_paths function}
\usage{
manage_paths(in_path = c(TRUE, FALSE), test_mode = c("all", "signalp2",
  "signalp3", "signalp4", "targetp", "tmhmm", "wolfpsort"), path_file = NULL)
}
\arguments{
\item{in_path}{if \strong{TRUE} manage_paths will attempt to run
external dependencies, assuming that respective paths are attached to the
$PATH variable;\cr
\cr
if \strong{FALSE} you should supply path_file.}

\item{test_mode}{if \strong{all} - all the external dependencies will be
checked;\cr
alternatively specify a tool name to be checked.}

\item{path_file}{full paths to external dependencies in a
2-column space-separated text file;\cr
for multiple versions of signalp please use 'signalpV'notation, where V is a
version number;\cr
\cr
\strong{first column} should contain tool name; \cr
\strong{second column} should contain full path to the tool's executable.}
}
\value{
a list of length 3 with the following elements:
\itemize{
\item \strong{tests}    TRUE if all the exteranl dependencies are working;
\item \strong{in_path}    TRUE if the dependecies are accessible via $PATH;
FALSE - if paths are provided with the path_file argument;
\item \strong{path_tibble} a tibble with verified names and paths for
external dependencies; \cr
NA if in_path is TRUE;\cr
if a specific tool is tested, a singualr path will be returned.
}
}
\description{
This function checks that all the external dependencies required for
secertome prediction pipelines are available and can sucessfully pass test
runs. \cr
\cr
All the external dependencies can be made accessible via $PATH environment
variable; alternatively a text file the with full paths to all the external
dependencies should be provided.
\cr
\cr
For the instalation instructions please see SecretSanta vignette.
}
\examples{
# Example1:
# here we assume that paths to all the
# external dependencies are attached to
# the $PATH variable:

manage_paths(in_path = TRUE, test_mode = 'all')

# to test just a single tool:
manage_paths(in_path = TRUE, test_mode = 'signalp2')

# Example2:
# alternatively, we are in a situation
# when changing $PATH is not possible,
# so we supply a file with listed full
# paths to the external dependencies:

manage_paths(in_path = FALSE,
test_mode = 'all',
path_file = system.file("extdata", "sample_paths", package = "SecretSanta"))
}
