% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_signalp.R
\name{parse_signalp}
\alias{parse_signalp}
\title{parse_signalp function}
\usage{
parse_signalp(some_param)
}
\arguments{
\item{param}{some param}
}
\description{
This function parses signalp2 and signalp3 output
}
\examples{
parse_signalp(some_param)
}
