% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived_functions.R, R/parse_signalP_web.R
\name{Parse_signalP_web}
\alias{Parse_signalP_web}
\alias{Parse_signalP_web}
\title{Parse_signalP_web function}
\usage{
Parse_signalP_web(url)

Parse_signalP_web(url)
}
\arguments{
\item{url}{url used to produce the output}

\item{url}{url used to produce the output}
}
\description{
This function scraps web output of SignalP tool and returns output in short format organised in a dataframe

This function scraps web output of SignalP tool and returns output in short format organised in a dataframe
}
\examples{
Parse_signalP_web()
}
\keyword{HMM,}
\keyword{peptide}
\keyword{peptide,}
\keyword{prediction,}
\keyword{proteins,}
\keyword{scrapping}
\keyword{signal}
\keyword{web}
