% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSP.R
\name{predictSP}
\alias{predictSP}
\title{PredictSP}
\usage{
predictSP(proteins)
}
\arguments{
\item{proteins}{input file/data frame with proteins}
}
\description{
This function takes input amino acid sequences and predict signla peptides in them
}
\examples{
predictSP()
}
\keyword{HMM}
\keyword{peptide}
\keyword{peptide,}
\keyword{prediction,}
\keyword{proteins,}
\keyword{signal}
