% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_signalp.R
\name{signalp}
\alias{signalp}
\title{signalp function}
\usage{
signalp(proteins, version, mode)
}
\arguments{
\item{proteins}{input file/data frame with proteins}

\item{version}{specify SignalP version to run}

\item{mode}{running mode, web or local}
}
\description{
This function calls local SignalP
}
\examples{
signalp()
}
\keyword{signalp}
