% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_signalp.R
\name{signalp}
\alias{signalp}
\title{signalp function}
\usage{
signalp(proteins, version, organism_type)
}
\arguments{
\item{proteins}{input file with proteins}

\item{version}{specify SignalP version to run; versions available: \cr
2.0 - http://www.cbs.dtu.dk/services/SignalP-2.0/ \cr
3.0 - http://www.cbs.dtu.dk/services/SignalP-3.0/ \cr
4.1 - http://www.cbs.dtu.dk/services/SignalP-4.1/ \cr}

\item{organism_type}{euk, gram+, gram-}
}
\description{
This function calls local SignalP
Please ensure that respective version of SignalP is downloaded, installed and respective path is added to $PATH variable.
For details see ... explain with aliases and examples
}
\examples{
result <- signalp(proteins = "SecretSanta/inst/extdata/sample_prot.fasta", organism_type = 'euk', version = 4)
}
