% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalp_parallel.R
\name{signalp}
\alias{signalp}
\title{signalp function}
\usage{
signalp(input_obj, version, organism = c("euk", "gram+", "gram-"),
  run_mode = c("starter", "piper"), paths = NULL, truncate = NULL,
  cores = NULL)
}
\arguments{
\item{input_obj}{an instance of CBSResult class containing protein 
sequences as one of the attributes}

\item{version}{signalp version to run, supported versions include: \cr
2, 3, 4.}

\item{organism}{\strong{euk} - for eukaryotes;\cr
\strong{gram+} - for gram-positive bacteria;\cr
\strong{gram-} - for gram-negative bacteria;\cr}

\item{run_mode}{\strong{starter} - if it is the first step in pipeline;\cr
\strong{piper} - if you run this function on the output of other CBS tools;}

\item{paths}{if required version of signalp is not acessible globally, a file
conatining a full path to it's executable should be provided; for details
please check SecretSanta vignette.}

\item{truncate}{if \strong{TRUE} - sequences longer 2000 residues will be 
truncated to this length limit and renamed;\cr
if \strong{FALSE} - long sequences will be excluded from the analysis;\cr
Default = TRUE.}

\item{cores}{optional arguments, number of cores to run the parallel process
on. If not set default will be 1.}
}
\value{
an object of SignalpResult class
}
\description{
This function calls local signalp to predict the presence and location of
signal peptide cleavage sites in amino acid sequences.
\cr
\cr
Large input files (>500 sequnces) are automatically split into smaller chunks
so that signalp prediction could be run as an embarassingly parallel process
on all the CPUs available.
}
\examples{

# read fasta file in AAStringSet object
aa <- readAAStringSet(system.file(
                       "extdata",
                       "sample_prot_100.fasta",
                        package = "SecretSanta"))
                                  
# assign this object to the input_fasta slot
# of empty CBSResult object
inp <- CBSResult(in_fasta = aa[1:10])

# run signalp2 on the initial file:
r1 <- signalp(inp,
              version = 2,
              organism = 'euk'
              run_mode = "starter")
}
