% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SignalpResult-class}
\alias{SignalpResult-class}
\alias{SignalpResult}
\title{An S4 class to represent intermediate and final outputs of the signalp prediction step}
\description{
An S4 class to represent intermediate and final outputs of the signalp prediction step
}
\section{Slots}{

\describe{
\item{\code{in_fasta}}{initial fasta file, do we need to drag it along?}

\item{\code{out_fasta}}{output fasta with only positive candidates, i.e sequences predicted to be secreted at this step}

\item{\code{mature_fasta}}{fasta with mature sequences}

\item{\code{sp_version}}{version of signalp used to generate this object}

\item{\code{sp_tibble}}{Object of class tibble, contains ... columns:
\itemize{
  \item gene_id - unique id of the sequence
  \item Cmax - max raw cleavage site score (C-score)
  \item Cpos - amino acid position with max C-score
  \item Ymax - max combied cleavage site score (Y-score)
  \item Ypos - amino acid position with max Y-score
  \item Smax - max signal peptide score
  \item Spos - amino acid position with max S-score 
  \item D - a weighted average of the mean S and the max. Y scores.
            This is the score that is used to discriminate signal peptides from non-signal peptides.
  \item Smean - the average S-score of the possible signal peptide (from position 1 to the position immediately before #              the maximal Y-score)
  \item Prediction - final desision o whether the protein is secreted or not (Y/N)
  \item Dmaxcut - 
  \item Networks used - 
  }}
}}

\examples{
a <- SignalpResult()
setInfasta(a, bb)
a2 <- setInfasta(a, bb)
getInfasta(a2)
}
