% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{methods}
\name{setMatfasta}
\alias{setMatfasta}
\alias{setMatfasta,SignalpResult-method}
\alias{setMatfasta}
\alias{getMatfasta,SignalpResult-method}
\alias{getMatfasta}
\alias{setSPversion,SignalpResult-method}
\alias{setSPversion}
\alias{getSPversion,SignalpResult-method}
\alias{getSPversion}
\alias{setSPtibble,SignalpResult-method}
\alias{setSPtibble}
\alias{getSPtibble,SignalpResult-method}
\alias{getSPtibble}
\title{Accessors for SignalpResult objects}
\usage{
setMatfasta(theObject, mature_fasta)

\S4method{setMatfasta}{SignalpResult}(theObject, mature_fasta)

\S4method{getMatfasta}{SignalpResult}(theObject)

\S4method{setSPversion}{SignalpResult}(theObject, sp_version)

\S4method{getSPversion}{SignalpResult}(theObject)

\S4method{setSPtibble}{SignalpResult}(theObject, sp_tibble)

\S4method{getSPtibble}{SignalpResult}(theObject)
}
\arguments{
\item{theObject}{SignalpResult object}

\item{mature_fasta}{seqeunces with clipped signal peptides, AAStringSet object}

\item{sp_version}{version of signalp used to generate SignalpResult object}

\item{sp_tibble}{parsed signalp output in tabular format}
}
\description{
Accessors for SignalpResult objects
}
