% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalp_parallel.R
\name{split_XStringSet}
\alias{split_XStringSet}
\title{split_XStringSet function}
\usage{
split_XStringSet(string_set, chunk_size)
}
\arguments{
\item{string_set}{- input AAStringSet that requires spliting;}

\item{chunk_size}{- number of sequenses in a single chunk;}
}
\description{
This function splits large XStringSet files into chunks of given size and returns a list of AAStringSets, those could be written to tmp files.
}
\examples{
aa <- readAAStringSet(system.file("extdata", 
                                   "sample_prot_100.fasta",
                                   package = "SecretSanta"))
split_XStringSet(aa,10)                                   
}
