% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_parallel.R
\name{split_XStringSet}
\alias{split_XStringSet}
\title{split_XStringSet function}
\usage{
split_XStringSet(string_set, chunk_size, residue_lim = NULL)
}
\arguments{
\item{string_set}{- input AAStringSet that requires spliting;}

\item{chunk_size}{- number of sequenses in a single chunk;}

\item{residue_lim}{- total number of residues/nucleotides allowed per chunk, if not met - smaller chunks are created
recursively for the offending chunk.}
}
\description{
This function splits large XStringSet files into chunks of given size and returns a list of AAStringSets, those could be written to tmp files.
}
\examples{
large_aa <- readAAStringSet(system.file("extdata", "Ppalm_prot_ALI_PLTG.fasta", package = "SecretSanta"))
split_XStringSet(large_aa, 1000)
res <- split_XStringSet(large_aa, 1000)
}
