% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalp_parallel.R
\name{split_XStringSet}
\alias{split_XStringSet}
\title{split_XStringSet function}
\usage{
split_XStringSet(string_set, chunk_size)
}
\arguments{
\item{string_set}{input AAStringSet object;}

\item{chunk_size}{the number of sequenses in a single chunk;}
}
\value{
list of AAStringSet chunks.
}
\description{
This function splits large XStringSet objects into chunks of given size and 
returns a list of AAStringSet objects.
}
\examples{
# Read fasta file:

aa <- readAAStringSet(
 system.file(
   "extdata", 
   "sample_prot_100.fasta",
    package = "SecretSanta"
 ))
 
# Split it into chunks
# with 10 sequences each: 
                                   
split_XStringSet(aa,10)                                   
}
