% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_targetP.R
\name{targetp}
\alias{targetp}
\title{tragetp function}
\usage{
targetp(proteins, output_type)
}
\arguments{
\item{proteins}{input file with proteins}

\item{output_type}{what we want from the output}
}
\description{
This function calls local targetp
Please check targetp README before running this fucntion:
TARGETP         'targetp-1.1' directory
TMP             location for temporary data
}
