% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetp_parallel.R
\name{targetp}
\alias{targetp}
\title{targetp function}
\usage{
targetp(input_obj, network, run_mode, paths)
}
\arguments{
\item{input_obj}{an instance of CBSResult class containing protein 
sequences in one of the slots}

\item{network}{\strong{P} - for plants; \cr
\strong{N} - for non-plants;}

\item{run_mode}{\strong{starter} - if it is the first step in pipeline; \cr
\strong{piper} - if you run this function on the output of other CBS tools;}

\item{paths}{if targetp is not acessible globally, a file
conatining a full path to it's executable should be provided; for details
please check SecretSanta vignette.}

\item{cores}{optional arguments, number of cores to run the parallel process
on. If not set default will be 1.}
}
\value{
an object of TargetpResult class
}
\description{
This function calls local targetp to predict subcellular localisation of 
a protein.\cr
\cr
For large inputs (\strong{>1000} sequences) will automatically run in
as a massive paralle job.
}
\examples{
my_pa <- manage_paths(system.file("extdata",
                                  "sample_paths",
                                  package = "SecretSanta"))
# read fasta file in AAStringSet object
aa <- readAAStringSet(system.file("extdata",
                      "sample_prot_100.fasta",
                      package = "SecretSanta"))
# assign this object to the input_fasta slot of SignalpResult object
inp <- CBSResult(in_fasta = aa[1:10])

tp_result <- targetp(input_obj = inp,
                     network = 'N',
                     run_mode = 'starter',
                     paths = my_pa)
}
