% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TargetpResult-class}
\alias{TargetpResult-class}
\alias{TargetpResult}
\title{An S4 class to represent intermediate and final outputs of the targetp prediction step}
\description{
An S4 class to represent intermediate and final outputs of the targetp prediction step
}
\section{Slots}{

\describe{
\item{\code{paths}}{paths to CBS exceutibles, do I need it here?}

\item{\code{in_fasta}}{initial fasta file, do we need to drag it along?}

\item{\code{out_fasta}}{output fasta with only positive candidates, i.e. not associated with cell organells}

\item{\code{tp_tibble}}{standard tibble with outputs obtained from targetp}
}}

