% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{methods}
\name{setTPtibble}
\alias{setTPtibble}
\alias{setTPtibble,TargetpResult-method}
\alias{setTPtibble}
\alias{getTPtibble}
\alias{getTPtibble,TargetpResult-method}
\alias{getTPtibble}
\title{Some title}
\usage{
setTPtibble(theObject, tp_tibble)

\S4method{setTPtibble}{TargetpResult}(theObject, tp_tibble)

getTPtibble(theObject)

\S4method{getTPtibble}{TargetpResult}(theObject)
}
\arguments{
\item{theObject}{some object}

\item{tp_tibble}{tp result}

\item{theObject}{some object}

\item{tp_tibble}{tp result}
}
\description{
Some title

Some title
}
