% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TargetpResult-class}
\alias{TargetpResult-class}
\alias{TargetpResult}
\alias{setTPtibble}
\alias{setTPtibble,TargetpResult-method}
\alias{getTPtibble,TargetpResult-method}
\alias{getTPtibble}
\title{accesor functions for objects of TargetpResult S4 class, intermediate and 
final outputs of the targetp prediction step}
\usage{
setTPtibble(theObject, tp_tibble)

\S4method{setTPtibble}{TargetpResult}(theObject, tp_tibble)

\S4method{getTPtibble}{TargetpResult}(theObject)
}
\arguments{
\item{theObject}{an object of TargetpResult class}

\item{tp_tibble}{parsed targetp output in tabular format}
}
\value{
TargetpResult object
}
\description{
accesor functions for objects of TargetpResult S4 class, intermediate and 
final outputs of the targetp prediction step

Accessors for TargetpResult objects
}
\section{Slots}{

\describe{
\item{\code{tp_tibble}}{tibble with outputs obtained from targetp
\itemize{
\item    gene_id - unique id of the sequence
\item    length - length of the protein sequence
\item    mTP - mitochondrial NN score
\item    sp - signal peptide NN score
\item    other - any onther NN score
\item    TP_localization - Prediction of localization, based on the scores;
the possible values are:
} 
\itemize{
\item    C - Chloroplast, i.e. the sequence contains cTP, a chloroplast
transit peptide;
\item    M - Mitochondrion, i.e. the sequence contains mTP, a mitochondrial
targeting peptide;
\item    S - Secretory pathway, i.e. the sequence contains SP, a signal
peptide;
\item    _ - Any other location;
\item    "don't know" - indicates that cutoff restrictions were set
(see instructions) and the winning network output score was below the 
requested cutoff for that category.
\item    RC - Reliability class, from 1 to 5, where 1 indicates the 
strongest prediction. RC is a measure of the size of the difference 
('diff') between #'  the highest (winning) and the second highest output
scores. There are 5 reliability classes, defined as follows:
\itemize{ 
    \item 1 - diff > 0.800;
    \item 2 - 0.800 > diff > 0.600
    \item 3 - 0.600 > diff > 0.400
    \item 4 - 0.400 > diff > 0.200
    \item 5 - 0.200 > diff
    }
}}
}}

\examples{
#read fasta file in AAStringSet object
aa <- readAAStringSet(system.file("extdata","sample_prot_100.fasta",
package = "SecretSanta"))
#assign this object to the input_fasta slot of SignalpResult object
inp <- CBSResult(in_fasta = aa[1:10])
tp_result <- targetp(input_obj = inp, network = 'N',
run_mode = 'starter')      
class(tp_result)
getTPtibble(tp_result)                     
getInfasta(tp_result)
getOutfasta(tp_result)
}
