% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmhmm.R
\name{tmhmm}
\alias{tmhmm}
\title{tmhmm function}
\usage{
tmhmm(input_obj, paths, TM)
}
\arguments{
\item{input_obj}{an instance of SignalpResult class, \cr
input should contain mature_fasta slot;}

\item{paths}{if tmhmm is not acessible globally, a file
conatining a full path to it's executable should be provided; for details
please check SecretSanta vignette.}

\item{TM}{allowed number of TM domains in mature peptides,
recommended value <= 1; use TM = 0 for strict filtering}
}
\value{
TMhmmResult object
}
\description{
This function calls local TMHMM to predict transmembarne domains in protein
sequence. \cr
\cr
In the full-length proteins N-terminal signal peptide could be
erroneously predicted as a TM domain. Therefore it is recommended to run
TMHMM on mature pepetide sequences (with clipped signal peptides) to avoid 
false positive predictions. \cr
\cr
To generate an CBSResult object with mature sequences please run signalp
predictions first.
}
\examples{
          
my_pa <- manage_paths(system.file("extdata", 
                      "sample_paths",
                      package = "SecretSanta"))
inp <- SignalpResult()
aa <- readAAStringSet(system.file("extdata",
                                  "sample_prot_100.fasta",
                                   package = "SecretSanta"))
inp <- setInfasta(inp, aa[1:10])
s1_sp2 <- signalp(inp,
                  version = 2,
                  'euk',
                  run_mode = "starter",
                  paths = my_pa)
tm <- tmhmm(s1_sp2, paths = my_pa, TM = 1)
}
