% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TMHMM.R
\name{tmhmm}
\alias{tmhmm}
\title{tmhmm function}
\usage{
tmhmm(input_obj)
}
\arguments{
\item{input_obj}{input object, an instance of CBSResult class, \cr
input should contain mature_fasta; in full-length proteins \cr
N-terminal signal peptide could be erroneously \cr
predicted as TM domain; avoid this}
}
\description{
This function calls local TMHMM
}
\examples{
r1 <- tmhmm("SecretSanta/inst/extdata/sample_prot.fasta", 'test')
}
