% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{methods}
\name{setTMtibble}
\alias{setTMtibble}
\alias{setTMtibble,TMhmmResult-method}
\alias{setTMtibble}
\alias{getTMtibble,TMhmmResult-method}
\alias{getTMtibble}
\alias{getInMatfasta,TMhmmResult-method}
\alias{getInMatfasta}
\alias{getOutMatfasta,TMhmmResult-method}
\alias{getOutMatfasta}
\title{Accessors for TMhmmResul objects}
\usage{
setTMtibble(theObject, tm_tibble)

\S4method{setTMtibble}{TMhmmResult}(theObject, tm_tibble)

\S4method{getTMtibble}{TMhmmResult}(theObject)

\S4method{getInMatfasta}{TMhmmResult}(theObject)

\S4method{getOutMatfasta}{TMhmmResult}(theObject)
}
\arguments{
\item{theObject}{object of TMhmmREsult class}

\item{tm_tibble}{parsed tmhmm output in tabular format}
}
\description{
Accessors for TMhmmResul objects
}
