% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{methods}
\name{setWOLFtibble}
\alias{setWOLFtibble}
\alias{setWOLFtibble,WolfResult-method}
\alias{setWOLFtibble}
\alias{getWOLFtibble,WolfResult-method}
\alias{getWOLFtibble}
\title{Accessors for WolfResult objects}
\usage{
setWOLFtibble(theObject, wolf_tibble)

\S4method{setWOLFtibble}{WolfResult}(theObject, wolf_tibble)

\S4method{getWOLFtibble}{WolfResult}(theObject)
}
\arguments{
\item{theObject}{an object of WolfResult class}

\item{wolf_tibble}{parsed output of wolfpsort in tabular format}
}
\description{
Accessors for WolfResult objects
}
