% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{WolfResult-class}
\alias{WolfResult-class}
\alias{WolfResult}
\alias{setWOLFtibble}
\alias{setWOLFtibble,WolfResult-method}
\alias{getWOLFtibble,WolfResult-method}
\alias{getWOLFtibble}
\title{accessor functions for objects of WolfResult S4 class, outputs of the wolf 
prediction step}
\usage{
setWOLFtibble(theObject, wolf_tibble)

\S4method{setWOLFtibble}{WolfResult}(theObject, wolf_tibble)

\S4method{getWOLFtibble}{WolfResult}(theObject)
}
\arguments{
\item{theObject}{an object of WolfResult class}

\item{wolf_tibble}{parsed output of wolfpsort in tabular format}
}
\value{
WolfResult object
}
\description{
accessor functions for objects of WolfResult S4 class, outputs of the wolf 
prediction step

Accessors for WolfResult objects
}
\section{Slots}{

\describe{
\item{\code{wolf_tibble}}{tibble with outputs obtained from wolfpsort
\itemize{
\item gene_id - unique sequence id
\item localization - the most probable localization
}}
}}

\examples{
aa <- readAAStringSet(system.file("extdata","sample_prot_100.fasta",
package = "SecretSanta"))
inp <- CBSResult(in_fasta = aa[1:10])
s1_sp2 <- signalp(inp, version = 2, organism = 'euk',
run_mode = "starter")
w <- wolfpsort(s1_sp2, 'fungi')
class(w)
#access result tibble:
getWOLFtibble(w)
#acess in_fasta and out_fasta slots:
getInfasta(w)
getOutfasta(w)
}
